Remove {Light,Vacuum}Interfaces
