+++ miio/integrations/mijia/vacuum/g1vacuum.py
@@ -1,11 +1,11 @@
 import logging
 from datetime import timedelta
 from enum import Enum
+from typing import Dict
 
 import click
 
 from miio.click_common import EnumType, command, format_output
-from miio.interfaces import FanspeedPresets, VacuumInterface
 from miio.miot_device import DeviceStatus, MiotDevice
 
 _LOGGER = logging.getLogger(__name__)
@@ -279,7 +279,7 @@ class G1CleaningSummary(DeviceStatus):
         return timedelta(hours=self.data["total_clean_area"])
 
 
-class G1Vacuum(MiotDevice, VacuumInterface):
+class G1Vacuum(MiotDevice):
     """Support for G1 vacuum (G1, mijia.vacuum.v2)."""
 
     _mappings = MIOT_MAPPING
@@ -379,7 +379,7 @@ class G1Vacuum(MiotDevice, VacuumInterface):
         return self.set_property("fan_speed", fan_speed.value)
 
     @command()
-    def fan_speed_presets(self) -> FanspeedPresets:
+    def fan_speed_presets(self) -> Dict[str, int]:
         """Return available fan speed presets."""
         return {x.name: x.value for x in G1FanSpeed}
