--- miio/interfaces/__init__.py
@@ -1,11 +1,0 @@
-"""Interfaces API."""
-
-from .lightinterface import ColorTemperatureRange, LightInterface
-from .vacuuminterface import FanspeedPresets, VacuumInterface
-
-__all__ = [
-    "FanspeedPresets",
-    "VacuumInterface",
-    "LightInterface",
-    "ColorTemperatureRange",
-]
